CI: Set caching based on the actual upstream compiler commit
The opam cache set up by `setup-ocaml` assumes that the compiler version
is enough to ensure that the proper compiler is in use, but this
assumption is not true for `trunk` compilers. So this adds an optional
key to the common workflow so that the `cache-prefix` key is either set
to v1 or to the hash of the current upstream compiler.

This also adds the benefit of logging the current commit id of the
compiler used by that CI run.