Add a print_vertical function
Sequential trace for STM tests were printed as a list, which is
inconstistent with the failure message printed just below them.

This commit proposes a print_vertical function on the model of
print_triple_vertical. This is the reason why the new function is placed
in the Util module, despite the fact that it is only used by STM.